修复点属性的显示和文件名显示

+++ libdxfrw/libdxfrw-master/src/drw_entities.h
@@ -838,7 +838,7 @@ public:
     }
 
     void update() {
-        numedges = objlist.size();
+        numedges = (int)objlist.size();
     }
 
 public:
